refactor: moving to route v1 and v2
