+++ svelte/src/routes/v1/lib/clean-canvas.ts
