+++ svelte/src/routes/v1/lib/graph-data/example-data.ts
