+++ svelte/src/routes/v1/lib/graph-data/index.ts
