+++ svelte/src/routes/v1/lib/graph-data/simple-data.ts
