+++ svelte/src/routes/v1/lib/setup-graph-data.ts
