+++ svelte/src/routes/v1/lib/style-injector.ts
