+++ svelte/src/routes/v2/components/draw-settings-changer.svelte
