+++ svelte/src/routes/v2/scripts/clean-canvas/index.ts
