+++ svelte/src/routes/v2/scripts/draw/drag-handler.ts
