+++ svelte/src/routes/v2/scripts/draw/index.ts
