+++ svelte/src/routes/v2/scripts/draw/tick.ts
